+++ NEWS.md
@@ -18,6 +18,7 @@
 * [SIP2-227](https://folio-org.atlassian.net/browse/SIP2-227): Resolve SSL connection failures when using PEM configuration
 * [SIP2-290](https://folio-org.atlassian.net/browse/SIP2-290): Upgrade module to Vert.x 5.0
 * [SIP2-287](https://folio-org.atlassian.net/browse/SIP2-287): Fix patron account blocking from unnecessary password verification
+* [SIP2-303](https://folio-org.atlassian.net/browse/SIP2-303): Fix SIP2 renewal response code for non-renewable loans
 
 
 ## 3.4.0 2025-03-14